+++ NutritionApp/app/src/main/java/com/bsnutrition/app/core/common/Dispatcher.kt
@@ -1,0 +1,14 @@
+package com.bsnutrition.app.core.common
+
+import javax.inject.Qualifier
+import kotlin.annotation.AnnotationRetention.RUNTIME
+
+@Qualifier
+@Retention(RUNTIME)
+annotation class Dispatcher(val bsnDispatcher: BsnDispatchers)
+
+enum class BsnDispatchers {
+    Default,
+    IO,
+    Main
+}